0.985 nightly

+++ Los Santos RED/lsr/UI/Pause Menu/Tabs/SaveGameTab.cs
@@ -69,54 +69,56 @@ public class SaveGameTab
             }
         };
         saveListItems.Add(createNew);
-
-        int maxNumber = GameSaves.GameSaveList.Max(x => x.SaveNumber);
-        for (int i = 1; i <= maxNumber; i++)
+        if (GameSaves.GameSaveList != null && GameSaves.GameSaveList.Any())
         {
-            UIMenuItem saveItem;
-            GameSave gs = GameSaves.GameSaveList.FirstOrDefault(x => x.SaveNumber == i);
-            if(gs != null)
+            int maxNumber = GameSaves.GameSaveList.Max(x => x.SaveNumber);
+            for (int i = 1; i <= maxNumber; i++)
             {
-                saveItem = new UIMenuItem(gs.Title, "") { RightLabel = gs.RightLabel };
-                if (GameSaves.IsPlaying(gs)) { saveItem.Text += " - Active"; saveItem.BackColor = EntryPoint.LSRedColor; }
-                saveItem.Activated += (s, e) =>
+                UIMenuItem saveItem;
+                GameSave gs = GameSaves.GameSaveList.FirstOrDefault(x => x.SaveNumber == i);
+                if (gs != null)
                 {
-                    SimpleWarning popUpWarning = new SimpleWarning("Save", "Are you sure you want to overwrite this save", "", Player.ButtonPrompts, Settings);
-                    popUpWarning.Show();
-                    if(popUpWarning.IsAccepted)
+                    saveItem = new UIMenuItem(gs.Title, "") { RightLabel = gs.RightLabel };
+                    if (GameSaves.IsPlaying(gs)) { saveItem.Text += " - Active"; saveItem.BackColor = EntryPoint.LSRedColor; }
+                    saveItem.Activated += (s, e) =>
                     {
-                        TabView.Visible = false;
-                        Game.IsPaused = false;
-                        int saveNumber = gs.SaveNumber;
-                        GameSaves.DeleteSave(gs);
-                        GameSaves.Save(Saveable, Weapons, Time, PlacesOfInterest, ModItems, saveNumber);
-                    }
-                };
-            }
-            else
-            {
-                saveItem = new UIMenuItem($"{i.ToString("D2")} - New Save Game", "");
-                saveItem.Activated += (s, e) =>
+                        SimpleWarning popUpWarning = new SimpleWarning("Save", "Are you sure you want to overwrite this save", "", Player.ButtonPrompts, Settings);
+                        popUpWarning.Show();
+                        if (popUpWarning.IsAccepted)
+                        {
+                            TabView.Visible = false;
+                            Game.IsPaused = false;
+                            int saveNumber = gs.SaveNumber;
+                            GameSaves.DeleteSave(gs);
+                            GameSaves.Save(Saveable, Weapons, Time, PlacesOfInterest, ModItems, saveNumber);
+                        }
+                    };
+                }
+                else
                 {
-                    int index = saveItem.Text.IndexOf(" -");
-                    string intAsString = (index > 0 ? saveItem.Text.Substring(0, index) : "");//why is this needed, is the i used as ref or something?
-                    int result = 0;
-                    int number = i;
-                    if (int.TryParse(intAsString, out result))
+                    saveItem = new UIMenuItem($"{i.ToString("D2")} - New Save Game", "");
+                    saveItem.Activated += (s, e) =>
                     {
-                        number = result;
-                    }
-                    SimpleWarning popUpWarning = new SimpleWarning("Save", "Are you sure you want to create a new save", "", Player.ButtonPrompts, Settings);
-                    popUpWarning.Show();
-                    if (popUpWarning.IsAccepted)
-                    {
-                        TabView.Visible = false;
-                        Game.IsPaused = false;
-                        GameSaves.Save(Saveable, Weapons, Time, PlacesOfInterest, ModItems, result);
-                    }
-                };
+                        int index = saveItem.Text.IndexOf(" -");
+                        string intAsString = (index > 0 ? saveItem.Text.Substring(0, index) : "");//why is this needed, is the i used as ref or something?
+                        int result = 0;
+                        int number = i;
+                        if (int.TryParse(intAsString, out result))
+                        {
+                            number = result;
+                        }
+                        SimpleWarning popUpWarning = new SimpleWarning("Save", "Are you sure you want to create a new save", "", Player.ButtonPrompts, Settings);
+                        popUpWarning.Show();
+                        if (popUpWarning.IsAccepted)
+                        {
+                            TabView.Visible = false;
+                            Game.IsPaused = false;
+                            GameSaves.Save(Saveable, Weapons, Time, PlacesOfInterest, ModItems, result);
+                        }
+                    };
+                }
+                saveListItems.Add(saveItem);
             }
-            saveListItems.Add(saveItem);
         }
         TabInteractiveListItem interactiveListItem2 = new TabInteractiveListItem("SAVE", saveListItems);
         TabView.AddTab(interactiveListItem2);
@@ -126,34 +128,36 @@ public class SaveGameTab
         List<UIMenuItem> saveListItems = new List<UIMenuItem>();
         UIMenuItem saveCount = new UIMenuItem($"Number of Save Games: {GameSaves.GameSaveList.Count()}", "") { Enabled = false};
         saveListItems.Add(saveCount);
-
-        int maxNumber = GameSaves.GameSaveList.Max(x => x.SaveNumber);
-        for (int i = 1; i <= maxNumber; i++)
+        if (GameSaves.GameSaveList != null && GameSaves.GameSaveList.Any())
         {
-            UIMenuItem loadItem;
-            GameSave gs = GameSaves.GameSaveList.FirstOrDefault(x => x.SaveNumber == i);
-            if (gs != null)
+            int maxNumber = GameSaves.GameSaveList.Max(x => x.SaveNumber);
+            for (int i = 1; i <= maxNumber; i++)
             {
-                loadItem = new UIMenuItem(gs.Title, "") { RightLabel = gs.RightLabel };
-                if (GameSaves.IsPlaying(gs)) { loadItem.Text += " - Active"; loadItem.BackColor = EntryPoint.LSRedColor; }
-                loadItem.Activated += (s, e) =>
+                UIMenuItem loadItem;
+                GameSave gs = GameSaves.GameSaveList.FirstOrDefault(x => x.SaveNumber == i);
+                if (gs != null)
                 {
-                    SimpleWarning popUpWarning = new SimpleWarning("Load", "Are you sure you want to load this save", "", Player.ButtonPrompts, Settings);
-                    popUpWarning.Show();
-                    if (popUpWarning.IsAccepted)
+                    loadItem = new UIMenuItem(gs.Title, "") { RightLabel = gs.RightLabel };
+                    if (GameSaves.IsPlaying(gs)) { loadItem.Text += " - Active"; loadItem.BackColor = EntryPoint.LSRedColor; }
+                    loadItem.Activated += (s, e) =>
                     {
-                        TabView.Visible = false;
-                        Game.IsPaused = false;
-                        GameFiber.Sleep(500);
-                        GameSaves.Load(gs, Weapons, PedSwap, Inventoryable, Settings, World, Gangs, Time, PlacesOfInterest, ModItems);
-                    }
-                };
+                        SimpleWarning popUpWarning = new SimpleWarning("Load", "Are you sure you want to load this save", "", Player.ButtonPrompts, Settings);
+                        popUpWarning.Show();
+                        if (popUpWarning.IsAccepted)
+                        {
+                            TabView.Visible = false;
+                            Game.IsPaused = false;
+                            GameFiber.Sleep(500);
+                            GameSaves.Load(gs, Weapons, PedSwap, Inventoryable, Settings, World, Gangs, Time, PlacesOfInterest, ModItems);
+                        }
+                    };
+                }
+                else
+                {
+                    loadItem = new UIMenuItem($"{i.ToString("D2")} - Empty Save", "") { Enabled = false };
+                }
+                saveListItems.Add(loadItem);
             }
-            else
-            {
-                loadItem = new UIMenuItem($"{i.ToString("D2")} - Empty Save", "") { Enabled = false };
-            }
-            saveListItems.Add(loadItem);
         }
         TabInteractiveListItem interactiveListItem2 = new TabInteractiveListItem("LOAD", saveListItems);
         TabView.AddTab(interactiveListItem2);
@@ -164,33 +168,35 @@ public class SaveGameTab
         List<UIMenuItem> saveListItems = new List<UIMenuItem>();
         UIMenuItem createNew = new UIMenuItem($"Number of Save Games: {GameSaves.GameSaveList.Count()}", "") { Enabled = false };
         saveListItems.Add(createNew);
-
-        int maxNumber = GameSaves.GameSaveList.Max(x => x.SaveNumber);
-        for (int i = 1; i <= maxNumber; i++)
+        if (GameSaves.GameSaveList != null && GameSaves.GameSaveList.Any())
         {
-            UIMenuItem deleteItem;
-            GameSave gs = GameSaves.GameSaveList.FirstOrDefault(x => x.SaveNumber == i);
-            if (gs != null)
+            int maxNumber = GameSaves.GameSaveList.Max(x => x.SaveNumber);
+            for (int i = 1; i <= maxNumber; i++)
             {
-                deleteItem = new UIMenuItem(gs.Title, "") { RightLabel = gs.RightLabel };
-                if (GameSaves.IsPlaying(gs)) { deleteItem.Text += " - Active"; deleteItem.BackColor = EntryPoint.LSRedColor; }
-                deleteItem.Activated += (s, e) =>
+                UIMenuItem deleteItem;
+                GameSave gs = GameSaves.GameSaveList.FirstOrDefault(x => x.SaveNumber == i);
+                if (gs != null)
                 {
-                    SimpleWarning popUpWarning = new SimpleWarning("Delete", "Are you sure you want to delete this save", "", Player.ButtonPrompts, Settings);
-                    popUpWarning.Show();
-                    if (popUpWarning.IsAccepted)
+                    deleteItem = new UIMenuItem(gs.Title, "") { RightLabel = gs.RightLabel };
+                    if (GameSaves.IsPlaying(gs)) { deleteItem.Text += " - Active"; deleteItem.BackColor = EntryPoint.LSRedColor; }
+                    deleteItem.Activated += (s, e) =>
                     {
-                        TabView.Visible = false;
-                        Game.IsPaused = false;
-                        GameSaves.DeleteSave(gs);
-                    }
-                };
+                        SimpleWarning popUpWarning = new SimpleWarning("Delete", "Are you sure you want to delete this save", "", Player.ButtonPrompts, Settings);
+                        popUpWarning.Show();
+                        if (popUpWarning.IsAccepted)
+                        {
+                            TabView.Visible = false;
+                            Game.IsPaused = false;
+                            GameSaves.DeleteSave(gs);
+                        }
+                    };
+                }
+                else
+                {
+                    deleteItem = new UIMenuItem($"{i.ToString("D2")} - Empty Save", "") { Enabled = false };
+                }
+                saveListItems.Add(deleteItem);
             }
-            else
-            {
-                deleteItem = new UIMenuItem($"{i.ToString("D2")} - Empty Save", "") { Enabled = false };
-            }
-            saveListItems.Add(deleteItem);
         }
         TabInteractiveListItem interactiveListItem2 = new TabInteractiveListItem("DELETE", saveListItems);
         TabView.AddTab(interactiveListItem2);
